Make ExcerptMaker main methods static
